Add playback controls and slider
