+++ features/media/src/main/res/drawable/ic_audio_icon_forward.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M12.1333,22C17.1298,22 21.2667,17.8631 21.2667,12.8667C21.2667,9.858 19.7445,7.1539 17.4432,5.4794C17.0492,5.175 16.5567,5.2735 16.3329,5.6406C16.109,6.0167 16.2254,6.4375 16.5925,6.7151C18.4998,8.0761 19.7355,10.3237 19.7445,12.8667C19.7534,17.0931 16.3597,20.4778 12.1333,20.4778C7.9069,20.4778 4.5312,17.0931 4.5312,12.8667C4.5312,9.2402 7.0205,6.2226 10.4052,5.4526L10.4052,6.742C10.4052,7.3777 10.8439,7.5479 11.3364,7.1986L14.1928,5.1929C14.5957,4.9153 14.6047,4.4855 14.1928,4.19L11.3453,2.1843C10.8439,1.8261 10.4052,1.9962 10.4052,2.6409L10.4052,3.9035C6.2146,4.7183 3,8.4612 3,12.8667C3,17.8631 7.1458,22 12.1333,22ZM15.0793,16.6364C16.5657,16.6364 17.5417,15.2306 17.5417,13.0816C17.5417,10.9147 16.5657,9.4909 15.0793,9.4909C13.5839,9.4909 12.6168,10.9057 12.6168,13.0816C12.6168,15.2306 13.5929,16.6364 15.0793,16.6364ZM9.3038,16.6364C10.7096,16.6364 11.7125,15.7858 11.7125,14.5769C11.7125,13.7442 11.1573,13.0547 10.3693,12.9114C11.0051,12.7324 11.4886,12.0697 11.4886,11.3176C11.4886,10.2431 10.6111,9.4909 9.3575,9.4909C8.274,9.4909 7.3697,10.0998 7.1816,10.9684C7.1548,11.0758 7.1458,11.1654 7.1458,11.2639C7.1458,11.5862 7.3518,11.8101 7.6562,11.8101C7.9159,11.8101 8.0771,11.6758 8.1756,11.3892C8.3546,10.8251 8.7397,10.5296 9.3306,10.5296C9.9574,10.5296 10.3872,10.9057 10.3872,11.4519C10.3872,12.025 9.9574,12.4458 9.3217,12.4458L8.9367,12.4458C8.6412,12.4458 8.4442,12.6518 8.4442,12.9652C8.4442,13.2607 8.6501,13.4845 8.9367,13.4845L9.3575,13.4845C10.1186,13.4845 10.5932,13.9054 10.5932,14.5322C10.5932,15.141 10.0559,15.6067 9.3127,15.6067C8.6859,15.6067 8.2382,15.2933 8.0412,14.7381C7.9338,14.4426 7.7726,14.3173 7.5309,14.3173C7.2264,14.3173 7.0115,14.5501 7.0115,14.8635C7.0115,14.962 7.0205,15.0605 7.0473,15.159C7.2443,16.0007 8.2024,16.6364 9.3038,16.6364ZM15.0793,15.5529C14.2913,15.5529 13.7809,14.6128 13.7809,13.0816C13.7809,11.5235 14.2823,10.5744 15.0793,10.5744C15.8762,10.5744 16.3687,11.5146 16.3687,13.0816C16.3687,14.6128 15.8672,15.5529 15.0793,15.5529Z"
+      android:strokeWidth="1"
+      android:fillColor="#000000"
+      android:fillType="nonZero"
+      android:strokeColor="#00000000"/>
+</vector>
